[FIX] Always send a specific NameIDFormat, depending on the integration type.
RealMe (MTS at least, possible ITE and Prod) will fail oddly if the NameIDFormat
provided in AuthN requests does not match the one supplied during metadata
upload/submission. On MTS, this results in the AuthN request being successfully
validated, but after submitting the resulting form, MTS fails with a HTTP 404
error rather than redirecting back to the issuer.

This simple change ensures that the NameIDFormat we use when running
dev/tasks/RealMeSetupTask matches the NameIDFormat that is used when submitting
the AuthN request.