FIX log in after logging in
When the config options for syncing members and logging a member in after authentication with RealMe are enabled, we should log a user in when they authenticate with RealMe. This was broken because a new RequestAuthenticationHandler contains no assigned member handlers, and thus the login() call was essentially a no-op. Using the named service AuthenticationHandler however (as defined in silverstripe/framework) has handlers assigned, and thus actually logs a member in after it has been matched to the RealMe FLT (or FIT) data.