DOCS update readme with relevant info for version 4 (#49)
Version 3 of this module worked with PHP 7.1 and still used the old mcrypt library.
The newer version 4 (yet to be released) no longer uses this fork, and sees a host
of issues disappear. The installation guide is now updated to reflect the new version,
along with the issue resolution that used to be required being moved to a troubleshooting
section, as it is undetermined (definitively) whether or not it is still required.
Testing with ITE has seen this fork no longer needed, however I am aware of a project
that does _seem_ to require it. So we'll leave it here as a note, rather than as a
specific instruction.