make css more specific as to avoid interference (#48)
The RealMe login detail box is strongly specified by the RealMe design team. It needs to look one of three pre-defiend ways, without modifications caused by cascading rules.
However the CSS supplied is not specific enough (nor perscriptive enough with the rules) to ensure that this is the case. So we can tighten this to be more robust.