FIX undefined variable dereference inside catch block
The auth variable is set inside the try to the result of a function call
however this function call may be the very place the exception that lands us
in the catch block is thrown. Throwing it twice instead of causing an
undefined variable error in PHP seems like a better way to let a developer
catch a clue (pun intended) as to what is going wrong.