FIX do not sync data with transient IDs
RealMe Assert type authentication is provided in order to allow users
to share some of their personal details with the service provider,
however this is provided with a transient NameID. That is, the shared
unique key used to identify a user between the IdP and the SP lasts
only for that singular session.

This means it is inappropriate for Silverstripe CMS to connect that
with a member object - every time the user logs in a new member object
will be created, and any data associated with it will be 'lost' from the
user's perspective the next time they log in.

Easiest solution: don't try to sync with local members. This can be done
by devlopers using another piece of shared information (such as an email
address), and does not need to be a part of this module.