FIX bad array index lookup for attributes
After successfully authenticating, the SAML response can have attributes
about the user's identity attached. A change was recently made to use
a newer JSON encoded attribute, instead of the XML that was traditionally
returned. However when logging in using the `login` type of RealMe
authentication, there is no identity information returned at all.

The old code accounted for this by inspecting to see if the XML attribute
existed, and skipped trying to retrieve the information if not.
The recent change however inspects if the XML attribute exists, only to
fall back to JSON instead - where this could also be validly missing.
When both are absent, this causes a PHP notice about bad array indexes.

Being the author responsible for making this logical mistake, I've now
corrected it. I've also updated the configuration documentation to make
it clearer that logging a user in after authentication is only possible
in `login` mode - not with `assert`.
Also corrected integration_type check before throwing exception for this
functionality to reference the correct type.